requirements.txt
asgiref==3.10.0
brotli==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
PyMySQL==1.1.2
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
whitenoise==6.11.0
